+++ database/migrations/2025_02_13_140503_add_table_mouvement.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('matiere_id')->constrained('matieres')->cascadeOnDelete();
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete(); // Qui a fait l'action
             $table->string('type'); // "entree" ou "sortie"
-            $table->integer('quantite');
-            $table->decimal('valeur_unitaire', 8, 3)->nullable();
+            $table->decimal('quantite', 16, 6);
+            $table->decimal('valeur_unitaire', 16, 6)->nullable();
             $table->string('raison')->nullable(); // Explication du mouvement
             $table->timestamp('date')->useCurrent();
             $table->foreignId('cde_ligne_id')->nullable()->constrained('cde_lignes')->onDelete('cascade');
